requirements.txt
requests
aiohttp
tenacity
flask
flask-cors
gunicorn
python-dotenv
psycopg2-binary
bcrypt
PyJWT
sendgrid  # 🚨 [신규] SendGrid 서비스 사용 시 필요
